+++ src/ppc/macro-assembler-ppc.cc
@@ -3972,7 +3972,7 @@ void MacroAssembler::LoadDoubleLiteral(DwVfpRegister result,
 }
 
 void MacroAssembler::Add(Register dst, Register src,
-                         uint32_t value, Register scratch) {
+                         intptr_t value, Register scratch) {
   if (is_int16(value)) {
     addi(dst, src, Operand(value));
   } else {
